Change the ids to long and change the update part route to work correctly

+++ src/java/controller/salePerson/part/DeletePartServlet.java
@@ -68,7 +68,7 @@ public class DeletePartServlet extends HttpServlet {
         
         try{
             PartsDAO partsDAO = new PartsDAO();
-            boolean hasDeleted = partsDAO.delete(Integer.parseInt(partId));
+            boolean hasDeleted = partsDAO.delete(Long.parseLong(partId));
             response.sendRedirect(request.getHeader("Referer"));
         }
         catch(Exception ex){
