+++ src/java/dao/PartsDAO.java
@@ -20,19 +20,19 @@ import mylib.DBUtils;
  * @author NGHIA
  */
 public class PartsDAO {
-    public  Parts getPartById(int id) throws ClassNotFoundException{
+    public  Parts getPartById(long id) throws ClassNotFoundException{
         String sqlQuery = "SELECT p.partID, p.partName, p.purchasePrice, p.retailPrice from Parts p WHERE p.partID = ?";
         Parts p = null;
         try{
             Connection con = DBUtils.getConnection();
             PreparedStatement ps = con.prepareStatement(sqlQuery);
             
-            ps.setInt(1, id);
+            ps.setLong(1, id);
             ResultSet rs = ps.executeQuery();
             if(rs!=null){
                 while(rs.next()){
                     p=new Parts();
-                    p.setPartID(rs.getInt("partID"));
+                    p.setPartID(rs.getLong("partID"));
                     p.setPartName(rs.getString("partName"));
                     p.setPurchasePrice(rs.getDouble("purchasePrice"));
                     p.setRetailPrice(rs.getDouble("retailPrice"));
@@ -69,7 +69,7 @@ public class PartsDAO {
                 list = new ArrayList<>();
                 while (rs.next()) {
                     Parts p = new Parts(
-                            rs.getInt("partID"),
+                            rs.getLong("partID"),
                             rs.getString("partName"),
                             rs.getDouble("purchasePrice"),
                             rs.getDouble("retailPrice")
@@ -100,7 +100,7 @@ public class PartsDAO {
         try {
             Connection con = DBUtils.getConnection();
             PreparedStatement ps = con.prepareStatement(sqlQuery);
-            ps.setInt(1, newParts.getPartID());
+            ps.setLong(1, newParts.getPartID());
             ps.setString(2, newParts.getPartName());
             ps.setDouble(3, newParts.getPurchasePrice());
             ps.setDouble(4, newParts.getRetailPrice());
@@ -127,7 +127,7 @@ public class PartsDAO {
             ps.setString(1, updatePart.getPartName());
             ps.setDouble(2, updatePart.getPurchasePrice());
             ps.setDouble(3, updatePart.getRetailPrice());
-            ps.setInt(4, updatePart.getPartID());
+            ps.setLong(4, updatePart.getPartID());
             int rows = ps.executeUpdate();
             
             status = rows>0;
@@ -142,7 +142,7 @@ public class PartsDAO {
         return (status)?updatePart:null;
     }
     
-    public  boolean delete(int id){
+    public  boolean delete(long id){
         boolean status = false;
         String deletePartsUsedSqlQuery = "DELETE FROM PartsUsed WHERE partID = ?";
         String deletePartsSqlQuery = "DELETE FROM Parts WHERE partID=?";
@@ -152,8 +152,8 @@ public class PartsDAO {
             PreparedStatement deletePartsUsePs = con.prepareStatement(deletePartsUsedSqlQuery);
             PreparedStatement deletePartsPs = con.prepareStatement(deletePartsSqlQuery);
 
-            deletePartsUsePs.setInt(1, id);
-            deletePartsPs.setInt(1, id);
+            deletePartsUsePs.setLong(1, id);
+            deletePartsPs.setLong(1, id);
             
             deletePartsUsePs.executeUpdate();
             status = deletePartsPs.executeUpdate()>0;
